Refactoring sonarqube bags from Reliability section. Issues: 1, 3, 6, 9, 16, 17, 18, 19, 20, 21, 22, 23, 24, 26, 29, 30, 31, 32, 33, 34

+++ carina-webdriver/src/main/java/com/qaprosoft/carina/core/foundation/webdriver/decorator/extractor/impl/DivisionElementExtractor.java
@@ -41,7 +41,7 @@ public class DivisionElementExtractor extends AbstractElementExtractor implement
         int isLower;
         Rectangle tempRect;
         while (elements.size() != 1) {
-            index = (int) (Math.round(elements.size() / 2));
+            index = Math.round(elements.size() / 2f);
             tempElement = elements.get(index);
             tempRect = getRect(tempElement);
             isLower = isLower(tempRect, y);
